requirements.txt
Django==2.2
django-taggit==1.4.0
flake8==3.9.0
Markdown==3.3.4
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.0
pytz==2021.1
sqlparse==0.4.1
yapf==0.31.0
